Implement version resolution
This commit implements version resolution using two tags and a single
internal function, specified in order in '#minecraft:load'. This
approach is not resilient to future changes to version resolution, but
as the process is so simple, there should be no changes necessary.

The implementation details (such as namespace, function / tag naming,
etc.) are subject to change. Ideally no more functions or tags would be
added though.