Make a few changes suggested by Aeldrion
This commit removes '"replace": false' because it's redundant and
apparently a significant amount of people think 'true' is a more
sensible default, so it's better to just not tempt them at all.

Additionally, the previous contents of '#minecraft:load' have been moved
into '#load:__init' to make sure all of Lantern Load's tags are loaded
in order, as a unit.