relicense to 0BSD license
TL;DR: CC0 -> 0BSD is basically meaningless for 99% of users.
The rationale for the relicense is contained for those who are curious.

The following text is not legal advice and I am not a laywer.

The new license is equally permissive to the old, but removes a ton of
fluff that Creative Commons had because of irrelevant concepts like
patents, public domain, and protecting their own Creative Commons
trademarks and branding. Public domain is actually legally problematic
in many jurisdictions, and the "fallback" clause in CC0 intended to
fix the problems is, to my knowledge, considered insufficient by many
corporations. For actual usage of software, whether it is public domain
or not is mostly irrelevant (there are cases where the distinction is
important, but Lantern Load is not one of those cases).

As CC0 was not built for software, it contains some questionable terms
for certain variants of software, primarily involving patents. Except
in jurisdictions where I successfully relinquished my copyright through
CC0 (notably, not the EU), I am still the copyright holder of most of
the software, so I have the universal right to relicense most of it.

The README was rewritten by another contributor. In many jurisdictions
the contribution is public domain so the relicense is irrelevant. In
jurisdictions where they are still the copyright holder, I am licensed
the right to the software under the CC0 fallback license, provided I
comply with the terms. The right to sublicense allows me to extend my
rights under a new license so long as it complies with all other terms.
The terms are basically nonexistent aside from patent law and Creative
Commons branding stuff, neither of which are relevant for this project.
Hence, I can freely sublicense the contribution under 0BSD, making the
whole repository successfully relicensed.